# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Runs Service",
    "description": "REST API for tracking service execution runs and their associated costs. Supports hierarchical runs (parent-child), cost aggregation, and multi-tenant isolation via organizations.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authenticating requests"
      }
    },
    "schemas": {
...
      "Run": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "brandId": {
            "type": "string",
            "nullable": true
          },
          "campaignId": {
            "type": "string",
            "nullable": true
          },
          "workflowName": {
            "type": "string",
            "nullable": true
          },
          "serviceName": {
            "type": "string"
          },
          "taskName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "parentRunId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "userId",
          "brandId",
          "campaignId",
          "workflowName",
          "serviceName",
          "taskName",
          "status",
          "parentRunId",
          "startedAt",
          "completedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "required": [
          "error"
        ]
      },
      "CreateRunRequest": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "minLength": 1,
            "deprecated": true,
            "description": "Deprecated: use x-brand-id header instead. Kept for backwards compatibility; header takes precedence."
          },
          "campaignId": {
            "type": "string",
            "minLength": 1,
            "deprecated": true,
            "description": "Deprecated: use x-campaign-id header instead. Kept for backwards compatibility; header takes precedence."
          },
          "workflowName": {
            "type": "string",
            "minLength": 1,
            "deprecated": true,
            "description": "Deprecated: use x-workflow-name header instead. Kept for backwards compatibility; header takes precedence."
          },
          "serviceName": {
            "type": "string",
            "minLength": 1
          },
          "taskName": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "serviceName",
          "taskName"
        ]
      },
      "ListRunsResponse": {
        "type": "object",
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunWithOwnCost"
            }
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "runs",
          "limit",
          "offset"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "AddCostsResponse": {
        "type": "object",
        "properties": {
          "costs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cost"
            }
          }
        },
        "required": [
          "costs"
        ]
      },
      "AddCostsRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostItem"
            },
            "minItems": 1
          }
        },
        "required": [
          "items"
        ]
      },
...
      "UpdateRunRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/runs": {
      "post": {
        "summary": "Create a run",
        "description": "Creates a new execution run. Organization and user are identified via x-org-id and x-user-id headers. Pass x-run-id header to set the parent run (the caller's run ID becomes parentRunId). Field resolution priority: header > body (deprecated) > parent inheritance. If a field conflicts with the parent run value, the request is rejected with 409. Returns 409 if orgId, userId, brandId, campaignId, or workflowName differ from the parent run.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Brand identifier, injected by workflow-service"
            },
            "required": false,
            "description": "Brand identifier, injected by workflow-service",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign identifier, injected by workflow-service"
            },
            "required": false,
            "description": "Campaign identifier, injected by workflow-service",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name, injected by workflow-service"
            },
            "required": false,
            "description": "Workflow name, injected by workflow-service",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Run created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or parentRunId does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Parent-child field conflict — request values differ from parent run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List runs",
        "description": "Lists runs for the organization identified by x-org-id header. Each run includes ownCostInUsdCents.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "userId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "workflowName",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "serviceName",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "taskName",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "parentRunId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "startedAfter",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "startedBefore",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of runs with cost totals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRunsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/runs/{id}": {
      "get": {
        "summary": "Get a run with costs",
        "description": "Returns the run with its cost breakdown, including all descendant runs and their costs",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Run with cost details and descendant runs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunWithCosts"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update run status",
        "description": "Updates the run status to completed or failed. Sets completedAt automatically.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{id}/costs": {
      "post": {
        "summary": "Add costs to a run",
        "description": "Adds cost line items. Unit costs are resolved automatically from the costs-service. Optional x-brand-id, x-campaign-id, x-workflow-name headers are forwarded to downstream services.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand identifier, injected by workflow-service"
            },
            "required": false,
            "description": "Brand identifier, injected by workflow-service",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign identifier, injected by workflow-service"
            },
            "required": false,
            "description": "Campaign identifier, injected by workflow-service",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name, injected by workflow-service"
            },
            "required": false,
            "description": "Workflow name, injected by workflow-service",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCostsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Costs added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddCostsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unknown cost name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{id}/costs/{costId}": {
      "patch": {
        "summary": "Update a cost item status",
        "description": "Updates a cost item status. Use to realize a provisioned cost (set status to 'actual') or cancel it (set status to 'cancelled').",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "costId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cost updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cost"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Run or cost not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}